staging: drm/omap: Create separate DMM driver
This patch splits the DMM off into a separate driver.  The DMM driver
utilizes the platform device information that was created as part of
the OMAP hwmod entry for the DMM.

The driver probe function queries the iomem resource and IRQ using
standard platform_get functions.

The OMAP DRM driver now calls the platform_driver_register() function
for the subordinate DMM driver inside its probe function.  This
guarantees the required order for the DMM and ensures the DMM resources
are available for use by the DRM driver.

TODO: Fix error case for no device, but cpu is omap44xx

Signed-off-by: Andy Gross <andy.gross@ti.com>