ARM: OMAP2+ mux: Correct the _omap_mux_init_gpio API.
The API always assumes that GPIO pins would be present only
in mode3 or mode4. But this is not true, they can be in any mux mode.
Just adding mode6 in to the API for OMAP5 now. But this has to
be corrected in a proper way to ensure that it is compatible for
all OMAP socs in general.

Signed-off-by: R Sricharan <r.sricharan@ti.com>